+++ UnnamedEngine/src/Platform/Linux/LinuxFilesystem.cpp
@@ -1,8 +1,13 @@
-#include "Core/Filesystem.hpp"
+#include "Core/PlatformIndependenceLayer/Filesystem.hpp"
 
 namespace UE
 {
-    UEPath Filesystem::m_InternalPath = "./env/Internal/";
-    UEPath Filesystem::m_AssetsPath = "./env/Assets/";
-    UEPath Filesystem::m_UserDataFolderPath = "";
+    UEPath Filesystem::s_InternalPath = "./env/Internal/";
+    UEPath Filesystem::s_AssetsPath   = "./env/Assets/";
+    UEPath Filesystem::s_UserDataPath = "";
+
+    UEResult<> Filesystem::SetUserDataPath()
+    {
+        return UEResult<>::Success;
+    }
 }