+++ UnnamedEngine/src/Source/Core/GlobalConfigurationSerializer.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "Core/Base.hpp"
+
+namespace UE
+{
+	class GlobalConfigurationSerializer
+	{
+	public:
+		static UEResult<> Serialize();
+		static UEResult<> Deserialize();
+	};
+}